requirements.txt
Django>=1.7.5,<1.8.0
djangorestframework>=3.0.5,<3.1.0
pytz
